[data] move many aggregate functions internal
