+++ python/ray/data/_internal/planner/exchange/aggregate_task_spec.py
@@ -1,10 +1,10 @@
 from typing import List, Optional, Tuple, Union
 
-from ray.data._internal.aggregate._aggregate import _AggregateOnKeyBase
+from ray.data._internal.aggregate import Count, _AggregateOnKeyBase
 from ray.data._internal.planner.exchange.interfaces import ExchangeTaskSpec
 from ray.data._internal.planner.exchange.sort_task_spec import SortKey
 from ray.data._internal.table_block import TableBlockAccessor
-from ray.data.aggregate import AggregateFn, Count
+from ray.data.aggregate import AggregateFn
 from ray.data.block import Block, BlockAccessor, BlockExecStats, BlockMetadata, KeyType
