+++ python/ray/data/preprocessors/discretizer.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 from ray.data import Dataset
-from ray.data._internal.aggregate._aggregate import Max, Min
+from ray.data._internal.aggregate import Max, Min
 from ray.data.preprocessor import Preprocessor
 from ray.util.annotations import PublicAPI
