+++ python/ray/data/preprocessors/imputer.py
@@ -6,7 +6,7 @@ import pandas as pd
 from pandas.api.types import is_categorical_dtype
 
 from ray.data import Dataset
-from ray.data.aggregate import Mean
+from ray.data._internal.aggregate import Mean
 from ray.data.preprocessor import Preprocessor
 from ray.util.annotations import PublicAPI
