+++ python/ray/data/tests/test_all_to_all.py
@@ -10,15 +10,7 @@ import pyarrow as pa
 import pytest
 
 import ray
-from ray.data._internal.aggregate._aggregate import (
-    Count,
-    Max,
-    Mean,
-    Min,
-    Quantile,
-    Std,
-    Sum,
-)
+from ray.data._internal.aggregate import Count, Max, Mean, Min, Quantile, Std, Sum
 from ray.data.aggregate import AggregateFn
 from ray.data.context import DataContext
 from ray.data.tests.conftest import *  # noqa
@@ -214,7 +206,7 @@ def test_groupby_large_udf_returns(ray_start_regular_shared):
 
 
 def test_agg_errors(ray_start_regular_shared):
-    from ray.data.aggregate import Max
+    from ray.data._internal.aggregate import Max
 
     ds = ray.data.range(100)
     ds.aggregate(Max("id"))  # OK
