+++ python/ray/data/tests/test_execution_optimizer.py
@@ -8,6 +8,7 @@ import pyarrow as pa
 import pytest
 
 import ray
+from ray.data._internal.aggregate import Count
 from ray.data._internal.execution.interfaces import ExecutionOptions
 from ray.data._internal.execution.operators.base_physical_operator import (
     AllToAllOperator,
@@ -55,7 +56,6 @@ from ray.data._internal.logical.util import (
 from ray.data._internal.planner.exchange.sort_task_spec import SortKey
 from ray.data._internal.planner.planner import Planner
 from ray.data._internal.stats import DatasetStats
-from ray.data.aggregate import Count
 from ray.data.context import DataContext
 from ray.data.datasource.parquet_datasink import _ParquetDatasink
 from ray.data.tests.conftest import *  # noqa
