+++ python/ray/train/tests/test_datasets_train.py
@@ -26,7 +26,7 @@ import torch.optim as optim
 import ray
 from ray import train
 from ray.air.integrations.mlflow import MLflowLoggerCallback
-from ray.data.aggregate import Mean, Std
+from ray.data._internal.aggregate import Mean, Std
 from ray.train import Checkpoint, DataConfig, RunConfig, ScalingConfig
 from ray.train.torch.torch_trainer import TorchTrainer
